requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
alembic
dotenv
nats-py